Handle compatibility branching in a compat module
